Strengthen pipeline robustness and finalize paper docs.
This update hardens feature/runtime reliability (deep metrics stability, required nonlinear dependencies, cross-cohort reporting, and test fixture validity) while adding reproducibility assets and complete manuscript sections including Table 2 prior-work comparison.

+++ fall_risk_pipeline/src/models/deep_models.py
@@ -525,7 +525,9 @@ class DeepModelTrainer:
             for Xb, _ in dl:
                 with self._autocast_context():
                     logits = model(Xb.to(self.device))
-                prob = torch.softmax(logits, dim=1).cpu().numpy()
+                # Force FP32 before softmax export to avoid AMP underflow/rounding
+                # that can produce invalid probability rows for sklearn AUC checks.
+                prob = torch.softmax(logits.float(), dim=1).cpu().numpy().astype(np.float32)
                 probs.append(prob)
         return np.vstack(probs)
